Homework-8-finished
Questions are in the mail, mainly concerning the toggler button - I have no idea why there is th blue border line, or bhow to make the hamburger menu more smooth, or why when I put there a picture with a design, only a part of the picture is displayed. Thanks for any comment. K.